Vadim_Oge_2026/olymp/network_anomaly_detector/requirements.txt
scapy==2.4.5
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.3
dash==2.14.1
plotly==5.15.0
tensorflow==2.13.0
joblib==1.3.2
PyYAML==6.0
requests==2.31.0
asyncio-mqtt==0.11.0